requirements.txt
aiogram~=2.19
python-dotenv~=0.20.0
tarantool~=0.7.1